Implement user roles and permission checking
Implemented user and admin roles and standard endpoints are now accessible to logged in users. I also implemented
a test admin endpoint which can only be accessed by admins and some functionalities (like deleting users or
updating users) can only be done by admins or the users themselves.

+++ src/main/java/ch/raising/controllers/AccountController.java
@@ -5,18 +5,24 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -33,6 +39,7 @@ import ch.raising.services.AccountService;
 import ch.raising.utils.JwtUtil;
 import ch.raising.controllers.AccountController;
 
+@RequestMapping("/account")
 @Controller
 public class AccountController {
 	@Autowired
@@ -58,7 +65,7 @@ public class AccountController {
 	 * @param account provided by the request 
 	 * @return response instance with message and status code
 	 */
-	@RequestMapping(value = "/account/login", method = RequestMethod.POST)
+	@RequestMapping(value = "/login", method = RequestMethod.POST)
 	@ResponseBody
 	public ResponseEntity<?> login(@RequestBody LoginRequest request) throws Exception {
         try {
@@ -67,7 +74,7 @@ public class AccountController {
         } catch (BadCredentialsException e) {
             throw new Exception("Incorrect username or password", e);
         }
-        
+
         final UserDetails userDetails = accountService.loadUserByUsername(request.getUsername());
         final String token = jwtUtil.generateToken(userDetails);
         
@@ -81,15 +88,16 @@ public class AccountController {
 	 */
 	@RequestMapping(value = "/register", method = RequestMethod.POST)
 	@ResponseBody
-	public ResponseEntity<Response> register(@RequestBody Account account) {
+	public ResponseEntity<?> register(@RequestBody Account account) {
 		return accountService.register(account);
 	}
 	
 	/**
-	 * Get all accounts
+	 * Get all accounts (only for admins)
 	 * @return list of all accounts
 	 */
-	@GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
+    @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
+    @PreAuthorize("hasRole('ROLE_ADMIN')")
 	@ResponseBody
 	public List<Account> getAccounts(){
 		return accountService.getAccounts();
@@ -98,22 +106,42 @@ public class AccountController {
 	/**
 	 * Searches for an account by id
 	 * @param id the id of the desired account
+     * @param request instance of the http request
 	 * @return details of specific account
 	 */
 	@GetMapping("/{id}")
 	@ResponseBody
-	public Account getAccountById(@PathVariable int id) {
-		return accountService.findById(id);
+	public ResponseEntity<?> getAccountById(@PathVariable int id, HttpServletRequest request) {
+        Account account = accountService.findById(id);
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+        boolean isAdmin = request.isUserInRole("ROLE_ADMIN");
+        if(account == null)
+            return ResponseEntity.status(500).body(new Response("Account not found"));
+
+        if(!account.getUsername().equals(username) && !isAdmin)
+            return ResponseEntity.status(403).body(new Response("Access denied"));
+
+		return ResponseEntity.ok().body(account);
 	}
 
 	/**
 	 * Delete account
 	 * @param id
+     * @param request instance of the http request
 	 * @return response object with status text
 	 */
 	@DeleteMapping("/{id}")
 	@ResponseBody
-	public ResponseEntity deleteAccount(@PathVariable int id) {
+	public ResponseEntity deleteAccount(@PathVariable int id, HttpServletRequest request) {
+        Account account = accountService.findById(id);
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+        boolean isAdmin = request.isUserInRole("ROLE_ADMIN");
+        if(account == null)
+            return ResponseEntity.status(500).body(new Response("Account not found"));
+
+        if(!account.getUsername().equals(username) && !isAdmin)
+            return ResponseEntity.status(403).body(new Response("Access denied"));
+
 		return accountService.deleteAccount(id);
-	}
+    }
 }
